feat: Add new columns to sport_leagues table
This commit adds new columns to the `sport_leagues` table in the database. Specifically, it adds the `sports_name` column, which will store the name of the sport associated with each league. This additional column will provide more flexibility and improve the organization of the data.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a clear and concise description of the changes made.